src/auth/requirements.txt
astroid==4.0.1
blinker==1.9.0
click==8.3.0
dill==0.4.0
Flask==3.1.2
Flask-MySQLdb==2.0.0
isort==7.0.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.3
mccabe==0.7.0
mysqlclient==2.2.7
parso==0.8.5
platformdirs==4.5.0
PyJWT==2.10.1
pylint==4.0.2
tomlkit==0.13.3
Werkzeug==3.1.3
bcrypt==4.0.1
jsonify==0.5
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
alembic==1.14.0